Merge prescription-notification-service into master

+++ AADWebApp/Models/BloodTestResult.cs
@@ -4,8 +4,8 @@ namespace AADWebApp.Models
 {
     public class BloodTestResult
     {
-        public int Id { get; set; }
-        public int BloodTestRequestId { get; set; }
+        public short Id { get; set; }
+        public short BloodTestRequestId { get; set; }
         public bool Result { get; set; }
         public DateTime ResultTime { get; set; }
     }
